% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{BettiNumbers}
\alias{BettiNumbers}
\title{Returns the  Betti numbers of a given mesh.}
\usage{
BettiNumbers(file_path)
}
\arguments{
\item{file_path}{A string: the path to a triangular mesh file in the ply format}
}
\description{
This function returns the Betti numbers b0, b1 and b2, which represents the number of connected components,
the number of independent tunnelsand the number of closed regions in space, respectively.
The function implementation followsthe algorithms described by Konkle, Moran, Hamann, and Joy in the work
Fast Methods for Computing Isosur-face Topology with Betti Numbers.
}
\examples{
BettiNumbers(system.file("extdata", "m3.ply", package = "tcie"))
}
