% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DataInfo}
\alias{DataInfo}
\title{Returns the dataset size and the isovalue range of a given dataset.}
\usage{
DataInfo(file_path)
}
\arguments{
\item{file_path}{A string: the path to the nhdr file}
}
\value{
A vector with the dataset size and the minimum and maximum values of a given dataset.
}
\description{
This function returns the size and isovalue range of a given dataset.
}
\examples{
DataInfo(system.file("extdata", "f3.nhdr", package = "tcie"))
DataInfo(system.file("extdata", "f3.nhdr", package = "tcie"))
}
